requirements.txt
Flask
python-docx
flask-cors
